Adding experiment for MaxRAM
Just execute `make run`, it will start a service with the following JVM arguments:
```
-XX:MinHeapSize=500M
-XX:MaxHeapSize=500M
-XX:MaxDirectMemorySize=500M
-XX:MaxMetaspaceSize=500M
-XX:MaxRAM=750M
```

After the service started (sorry for the sleep command), it will ask the service to reserve 450MB heap and 450MB off-heap memory. Both should be successbul and at the end you should see that the total memory consumption is over a GiB but definitey over 900MiB (450+450).

This means that `-XX:MaxRAM` does not limit the total memory consumption of the JVM  but it serves as a base for the JVM Ergonomics: https://docs.oracle.com/en/java/javase/11/gctuning/ergonomics.html